Add a `hasImagesPendingUpload` extension helper to check if there are still any images being uploaded. Add unit tests.
